+++ src/components/matchup-planner/matchup-planner-widget/MatchupPlannerWidget.scss
@@ -20,31 +20,3 @@
     height: 80vh;
   }
 }
-
-.warlord-list {
-  height: 100%;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.warlord-card {
-  width: 315px;
-  height: 135px;
-
-  .inner {
-    background-size: contain;
-    background-repeat: no-repeat;
-    background-color: #3d2419;
-    margin: 5px;
-    background-position: center;
-    width: 100%;
-    height: 100%;
-  }
-
-  .delete-button {
-    float: right;
-    background-color: red;
-    padding: 3px;
-    margin: 5px;
-  }
-}
