+++ src/components/matchup-planner/warlord-card/WarlordCard.scss
@@ -1,0 +1,21 @@
+.warlord-card {
+  width: 315px;
+  height: 135px;
+
+  .inner {
+    background-size: contain;
+    background-repeat: no-repeat;
+    background-color: #3d2419;
+    margin: 5px;
+    background-position: center;
+    width: 100%;
+    height: 100%;
+  }
+
+  .delete-button {
+    float: right;
+    background-color: red;
+    padding: 3px;
+    margin: 5px;
+  }
+}
