+++ src/components/matchup-planner/warlord-list/WarlordList.scss
@@ -1,0 +1,5 @@
+.warlord-list {
+  height: 100%;
+  overflow-y: auto;
+  overflow-x: hidden;
+}
